PIFO pipeline reference model, which models:
1. A single "PIFO".
2. A "priority queue" / "calendar queue" wrapped over the PIFO.
3. A "stage" consisting of many "priority queues" and/or "calendar queues".
4. On the input side, the stage consists of a priority "computer"
that computes priorities/departure times before enqueuing into the
priority/calendar queues in that stage.
5. On the output side, the stage consists of a look-up table that
directs dequeue results to the next hop for further processing.

Based on Da's suggestion, we are assuming the entire pipeline executes
in one clock cycle end-to-end (as opposed to one clock cycle per
pipeline stage). But, we'll be imposing constraints on the number of ops
per clock within each stage of this pipeline.  Runs now, haven't
verified the results though.  Moving onto generating reasonable tests.